# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Project API",
    "description": "API for managing projects, daily details, and monthly reports with MSSQL integration for a Flutter calendar app",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Authenticate user and generate JWT token using username only",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username"],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "User's username"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login, returns JWT token and user details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "JWT token for authentication"
                    },
                    "userId": {
                      "type": "integer",
                      "description": "User's ID"
                    },
                    "Username": {
                      "type": "string",
                      "description": "User's username"
                    },
                    "Role": {
                      "type": "string",
                      "description": "User's role (user, group_manager, general_manager, finance_manager)"
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "401": { "description": "Invalid username" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/monthly-reports/{year}/{month}": {
      "post": {
        "summary": "Create monthly report (by user, group_manager, or general_manager)",
        "tags": ["MonthlyReports"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "in": "path",
            "name": "year",
            "required": true,
            "schema": { "type": "integer" }
          },
          {
            "in": "path",
            "name": "month",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "201": { "description": "Report created" },
          "400": { "description": "Invalid input" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/monthly-reports/{reportId}/submit-to-group-manager": {
      "put": {
        "summary": "Submit report to group manager (by user or group manager)",
        "tags": ["MonthlyReports"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "in": "path",
            "name": "reportId",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Submitted to group manager" },
          "403": { "description": "Access denied" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/monthly-reports/{reportId}/approve-group-manager": {
      "put": {
        "summary": "Approve and submit to general manager or finance (by group manager)",
        "tags": ["MonthlyReports"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "in": "path",
            "name": "reportId",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "comment": { "type": "string" },
                  "toGeneralManager": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Approved and submitted" },
          "403": { "description": "Access denied" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/monthly-reports/{reportId}/approve-general-manager": {
      "put": {
        "summary": "Approve and submit to finance (by general manager)",
        "tags": ["MonthlyReports"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "in": "path",
            "name": "reportId",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "comment": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Approved and submitted to finance" },
          "403": { "description": "Access denied" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/monthly-reports/{reportId}/approve-finance": {
      "put": {
        "summary": "Final approve (by finance)",
        "tags": ["MonthlyReports"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "in": "path",
            "name": "reportId",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "comment": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Approved" },
          "403": { "description": "Access denied" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/monthly-reports/{reportId}": {
      "get": {
        "summary": "Get report by ID",
        "tags": ["MonthlyReports"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "in": "path",
            "name": "reportId",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Report details" },
          "403": { "description": "Access denied" },
          "404": { "description": "Report not found" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/monthly-reports/group/{year}/{month}": {
      "get": {
        "summary": "Get reports for group manager or general manager",
        "tags": ["MonthlyReports"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "in": "path",
            "name": "year",
            "required": true,
            "schema": { "type": "integer" }
          },
          {
            "in": "path",
            "name": "month",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "List of reports" },
          "403": { "description": "Access denied" },
          "500": { "description": "Server error" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MonthlyReport": {
        "type": "object",
        "properties": {
          "ReportId": { "type": "integer", "description": "Unique identifier for the report" },
          "UserId": { "type": "integer", "description": "Identifier of the user" },
          "Year": { "type": "integer", "description": "Year of the report" },
          "Month": { "type": "integer", "description": "Month of the report (1-12)" },
          "TotalHours": { "type": "integer", "description": "Total hours worked in the month" },
          "GymCost": { "type": "integer", "description": "Gym cost for the month" },
          "Status": { "type": "string", "description": "Status of the report (draft, submitted_to_group_manager, submitted_to_general_manager, submitted_to_finance, approved)" },
          "ManagerComment": { "type": "string", "description": "Comment from the manager" },
          "FinanceComment": { "type": "string", "description": "Comment from the finance team" },
          "SubmittedAt": { "type": "string", "format": "date-time", "description": "Date and time when the report was submitted to manager" },
          "ApprovedAt": { "type": "string", "format": "date-time", "description": "Date and time when the report was final approved" },
          "GroupId": { "type": "integer", "description": "Identifier of the group" },
          "GeneralManagerStatus": { "type": "string", "description": "Status of general manager approval (pending, approved_by_general_manager)" }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}